https://github.com/OpenPaaS-Suite/esn-frontend-inbox/issues/53: Moved inboxEmail hook services to esn.inbox.libs
