+++ src/esn.inbox.libs/app/services/hook/email-composing-hook.service.spec.js
